discord bot.py/requirements.txt
# Discord Bot Requirements

# Core Discord library
discord.py>=2.3.0

# Database
aiosqlite>=0.19.0

# Environment variables
python-dotenv>=1.0.0

# System monitoring (for health check)
psutil>=5.9.0

# Optional: Genshin Impact integration
genshin>=1.7.0
